Get rid off all signed vs. unsigned compile warnings.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10728 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/configuration/AudioConfigurationPage.cpp
@@ -254,7 +254,7 @@ AudioConfigurationPage::getBestAvailableAudioEditor()
 
         QString n(candidates[i]);
 
-        for (size_t j = 0;
+        for (int j = 0;
              j < pathList.size() && result == "";
              j++) {
