+++ src/gui/editors/segment/compositionview/SegmentMover.cpp
@@ -171,7 +171,7 @@ void SegmentMover::handleMouseButtonRelease(QMouseEvent *e)
 
                 if (trackPos < 0) {
                     trackPos = 0;
-                } else if (trackPos >= comp.getNbTracks()) {
+                } else if (trackPos >= (int)comp.getNbTracks()) {
                     trackPos = comp.getNbTracks() - 1;
                 }
 
@@ -270,7 +270,7 @@ int SegmentMover::handleMouseMove(QMouseEvent *e)
 
         if (trackPos < 0) {
             trackPos = 0;
-        } else if (trackPos >= comp.getNbTracks()) {
+        } else if (trackPos >= (int)comp.getNbTracks()) {
             trackPos = comp.getNbTracks() - 1;
         }
 /*!!!
